Move function to encoding file
